+++ .gitignore
@@ -1,4 +1,3 @@
 target/
 .idea/
 *.iml
-ru.levelup.kseniia.alekseeva.qa.homework_2/*
