+++ src/main/java/ru/levelup/kseniia/alekseeva/qa/homework_2/Airline/organisational/Airline.java
@@ -2,7 +2,7 @@ package ru.levelup.kseniia.alekseeva.qa.homework_2.Airline.organisational;
 
 import ru.levelup.kseniia.alekseeva.qa.homework_2.Airline.transport.aircrafts.Plane;
 import java.util.TreeSet;
-/*
+/**
  * This is an Airline class that has method findPlanesInRangeByFlightHours() where we can find a plane that lies within specified range of flying hours
  *
  */
@@ -14,8 +14,12 @@ public class Airline {
         return airlineCarryingCapacity;
     }
 
-    public void setAirlineCarryingCapacity(int airlineCarryingCapacity) {
-        this.airlineCarryingCapacity = airlineCarryingCapacity;
+    public void setAirlineCarryingCapacity(TreeSet<Plane> planes) {
+        int temp = 0;
+        for (Plane plane : planes) {
+            temp = temp + plane.getCarryingCapacity();
+        }
+        this.airlineCarryingCapacity = temp;
     }
 
     private int airlineCarryingCapacity;
@@ -24,8 +28,12 @@ public class Airline {
         return seatsNumberInAllAircrafts;
     }
 
-    public void setSeatsNumberInAllAircrafts(int seatsNumberInAllAircrafts) {
-        this.seatsNumberInAllAircrafts = seatsNumberInAllAircrafts;
+    public void setSeatsNumberInAllAircrafts(TreeSet<Plane> planes) {
+        int temp = 0;
+        for (Plane plane : planes) {
+            temp = temp + plane.getSeatsNumber();
+        }
+        this.seatsNumberInAllAircrafts = temp;
     }
 
     private int seatsNumberInAllAircrafts;
@@ -37,19 +45,7 @@ public class Airline {
         this.pilotsNumber = pilotsNumber;
     }
 
-    public void findPlanesInRangeByFlightHours(TreeSet<Plane> planes, int from, int to) {
-        System.out.println("Planes, which have flight hours from " + from + " to " + to);
-        int number = 0;
-        for (Plane plane : planes) {
-            if (plane.getFlightHours() > from && plane.getFlightHours() < to) {
-                System.out.println(plane.toString() + ", flight hours " + plane.getFlightHours());
-                number++;
-            }
-        }
-        if (number == 0) {
-            System.out.println("Planes, which have flight hours from " + from + " to " + to + " not found!");
-        }
-    }
+
 
 
 }
