+++ src/main/java/ru/levelup/kseniia/alekseeva/qa/homework_2/Airline/organisational/Main.java
@@ -7,7 +7,7 @@ import ru.levelup.kseniia.alekseeva.qa.homework_2.Airline.transport.aircrafts.Pl
 
 import java.util.TreeSet;
 
-/*
+/**
  * This is the Main class where we create instances of AirbusA319, AirbusA320, Boeing737
  * we create a TreeSet of planes sorted by flying range
  * we find planes that match following range of flight hours: from 100 to 400
@@ -38,26 +38,28 @@ public class Main {
         boeing737.setCarryingCapacity(2500);
         boeing737.setName("boeing737");
 
-
-        fatCat.setSeatsNumberInAllAircrafts(airbusA319.getSeatsNumber() + airbusA320.getSeatsNumber() + boeing737.getSeatsNumber());
-        System.out.println("Number of seats in all aircrafts of airline 'Fat Cat' = " + fatCat.getSeatsNumberInAllAircrafts());
-
-        fatCat.setAirlineCarryingCapacity(airbusA319.getCarryingCapacity() + airbusA320.getCarryingCapacity() + boeing737.getCarryingCapacity());
-        System.out.println("Carrying capacity of airline 'Fat Cat' = " + fatCat.getAirlineCarryingCapacity() + " t");
-
         TreeSet<Plane> planes = new TreeSet<>();
         planes.add(airbusA319);
         planes.add(airbusA320);
         planes.add(boeing737);
+
+        fatCat.setSeatsNumberInAllAircrafts(planes);
+        System.out.println("Number of seats in all aircrafts of airline 'Fat Cat' = " + fatCat.getSeatsNumberInAllAircrafts());
+
+        fatCat.setAirlineCarryingCapacity(planes);
+        System.out.println("Carrying capacity of airline 'Fat Cat' = " + fatCat.getAirlineCarryingCapacity() + " t");
 
         System.out.println("List of planes, sorted by flying range:");
         for (Plane plane : planes) {
             System.out.println(plane.toString() + ", Flying range " + plane.getFlyingRange());
         }
 
-        fatCat.findPlanesInRangeByFlightHours(planes, 100, 400);
+        Search search = new Search();
+        search.findPlanesInRangeByFlightHours(planes, 100, 400);
 
     }
+
+
 
 
 }
