+++ src/main/java/ru/levelup/kseniia/alekseeva/qa/homework_2/Airline/transport/aircrafts/Airbuses/Airbus.java
@@ -2,7 +2,7 @@ package ru.levelup.kseniia.alekseeva.qa.homework_2.Airline.transport.aircrafts.A
 
 import ru.levelup.kseniia.alekseeva.qa.homework_2.Airline.transport.aircrafts.Plane;
 
-/*
+/**
  * This is Airbus class that has two methods 'autopilot' and 'touchdown' that will be used in methods fly () and land () for child classes
  */
 public class Airbus extends Plane {
