+++ src/main/java/ru/levelup/kseniia/alekseeva/qa/homework_2/Airline/transport/aircrafts/Airbuses/AirbusA320.java
@@ -3,7 +3,7 @@ package ru.levelup.kseniia.alekseeva.qa.homework_2.Airline.transport.aircrafts.A
 import ru.levelup.kseniia.alekseeva.qa.homework_2.Airline.transport.aircrafts.Flying;
 import ru.levelup.kseniia.alekseeva.qa.homework_2.Airline.transport.aircrafts.Maintainable;
 
-/*
+/**
  * This is an AirbusA320 class that implements three methods from Flying takeoff(); fly(); land ();
  * it also implements two methods from Maintainable isPreparedForDeparture() and passQuaterlyMaintenance()
  * AirbusA320 amd AirbusA319 have similar realization for fly(); land (); but different implementation for takeoff();
@@ -11,16 +11,7 @@ import ru.levelup.kseniia.alekseeva.qa.homework_2.Airline.transport.aircrafts.Ma
 public class AirbusA320 extends Airbus implements Maintainable, Flying {
 
     public AirbusA320() {
-        System.out.println("This is how AirbusA320 takes off");
-        takeoff();
-        System.out.println("This is how AirbusA320 flies");
-        fly();
-        System.out.println("This is how AirbusA320 lands");
-        land();
-        System.out.println("This is how AirbusA320 is prepared for departure");
-        isPreparedForDeparture();
-        System.out.println("This is how AirbusA320 passes quaterly maintenance");
-        passQuaterlyMaintenance();
+
     }
 
     public void takeoff() {
@@ -40,7 +31,9 @@ public class AirbusA320 extends Airbus implements Maintainable, Flying {
         System.out.println("Execute Check for 'Airbus A320'");
         return true;
     }
-
+    /**
+     The AirbusA320 passes examinations depending on the flight hours number
+     */
     public void passQuaterlyMaintenance() {
         if (this.getFlightHours() > 750 && this.getFlightHours() <= 8600) {
             System.out.println("Execute A Check");
